CHANGES TO THE SOCIAL CHOICE SIM
So now, the previous votes are passed in and then all the votes are collected at the end. This allows me to keep track of all of the votes a little more easily, as well as allowing for full cycle votes to affect everything. This might be expanded even further as we need to give more information to the bots (especially relating to social stuff) but! unitl then this should be it. Let me know if you have any quesitons.