fixed the 5 bug, other
made a lot of changes to teh way that data is passed into the sim, so that might be a little confusing, but basically just pass in the round and the cycle when graphing, and then when actually starting the sim pass in the situations and hwatnot. I don't know if it will work the way that I intend it to for human players as well, but thats ok for now. I will talk to garrett later about implementing more human testing, for now I am just happy that we are seeing some interesting behavrion.
Specifically, I would refer you to round 1 situation limitedawareGreedy+secondChoice, cycles 1 and 2. there is a clear swap from player 7 from cuase 1 to cause 3. unfortunatley I don't see a cause passing regardless, 3 might have passed if hte seocnd choice fetchers were distribuited evenly ,but after accounting for that 1 just had more peeps. very cool. more work to be donie but I am encouraged by the current progress.