added comparison software
so first I added a way to create all the graphs explicity from a single json, which I will be using for human testing, and then i created some comparison software to compare human and bot results. Be warned! it is not yet as complete as I would like it to be. But! we are getting there. Most of the offline sim stuff will be getting moved to the main branch in time, just becuase I don't want to stay on this branch for longer than I have too. Cheers! don't look too closely at it.