we had to do some more axe$ mods
but hey it works now. I hadn't sorted out the logic between the top half of the circle and bottom half of the circle, and I had to go back and fix that. But! finally! it works. I'm not gonna try and refactor or change it, it works the way its needs to and I am done with it.