they have been separated
there is still a lot of work to do, but obstenibly some work has been made. hte JHG client is no longer completely broken, but getting the rest of the button infrastructure to work the way that I want it to is still a work in progress. lets get it.