the amounts are now toggleable
fantstic. in the otpinso menu of the server, you can now set how many tokens per playe rin JHG and utilty per player in the allocations portion of SC and it all works offline and online. fantastic. I need to do some serious refactoring of the client likely if we want to continue forward with this.