Added a logger, got social bots working, graphing
polished the complete logger, got the batch system more or less working and tested it by rounding out some simple social welfare bots to set a baseline for us as we move forward. there are more steps to take tomorrow but for now lets just get this polished. Have a list of changes in my keep that I need to make.