BOOM
got the differing amounts of JHG and SC rounds going. debating if there is a scenario in which I would ever consider runnign like 2 sc games back to back, but I don't really see it. baking it into the sim would also be a pain so imma leave it alone for now.