GENETIC ALGORITHM NOW WORKS
python implementation rough draft is now complete for the jhg game. I have decided that the next step is understanding the gene3 bot better so I can adapt its current implementation into a bot that can also vote. Is this a little overkill? yeah, but whatever. Using this same bot in a variety of test beds is the whole point, so some basic or even minimal changes should be the highest.