calling autoscale finally seemed to fix it
so like, Idk what in those final changes finally fixed it but it seems to work now. subject to more rigourous testing when I walk in tomorrow but that finally seems to update when its ready rather than being forced by an outside source, even if it takes a while.