fixed the off by one error
finally stripped out the idea of receiving winning vote from the server, especially if we just get all the votes. basing it off of the already stored votes was less work right now. Its not as elegnat of a solution but it does work, which is what I get paid to do.