graphs, headless, logger
deleted a bunch of graphs, made the server headless if we really wanted to (turns out you can just play with no one if you really want to, makes streamlining data collection WAAY easier) and with that aforementioned headless mode it was much easier to figure out where I had messed up with the logger in the server version. all in a day's work.