added availability to both pure sc and pur3e jhg
changed the way the logger and rounds system worked to allow the toggleation of either 1 or the other within the fetcher, which should allow for better and more isolated preformance. tomorrow we need to tackle how to handle the difference between recieved and the influence matricies to see the best way to translate those over.