that bug was WEIRD
ok so, trying to run the anti cat bot training program. Kept getting an issue that the community was not a set. Tried to just throw that out anyway and hard convert it. it broke in ways that I have never seen before. python didn't even throw an error, just terminated early. Never figured out why. Undid the thing, mucked around for a while. I totally forgot that in the SC sim I changed that I no longer use the curr_sc_round in a way that matters, and instead just rely on nomral round usage. this was great - but it was resetting the bots on teh back end, and causing them to use a list for their thingies, which was beyond weird. Fixed the round issue and now it appears to be running as intedned (at least, it doesn't crash). further testing required to see if its actually doing what I want to do or not.