found the problem
promise not to kill me
I changed the num_gene_pools on the other thing to 1 instead of 3
leading to inconsistent behavior
changing that back reveals that there is no problem with the engine, it is perfectly healthy!
so instead its a problem with the training pipeline
so I will change that later.