homo and mixed algorithms done
created way to freeze seed and was able to prove that results between teh two are consistent, then I refactored the entire genetic library to be neater. we will run a bunch of stuff with the new and old cats now and also in self play to try and produce new kinds of genes for these bots.